Move ChatObservableImpl to separate file
